♻️ (server): Move operation usecase to its domain
